Fix warning: Incompatible pointer types passing 'NSUInteger *' to parameter of type 'unsigned int *'

+++ NSObject-AutoDescribe/NSObject+AutoDescribe.m
@@ -15,7 +15,7 @@
 {
     id instance = self;
     
-    NSUInteger count;
+    unsigned int count;
     objc_property_t *propList = class_copyPropertyList(classType, &count);
     NSMutableString *propPrint = [NSMutableString string];
